changed to int to let compiler choose most efficient int to use

+++ array-chunking/main.go
@@ -3,6 +3,6 @@ package main
 func main() {
 }
 
-func chunk(a []int8, size int8) (result [][]int8) {
+func chunk(a []int, size int) (result [][]int) {
 
 }
